feat: refactor MediaPage to use MediaInner component for improved structure and maintainability
